Rename files jsx -> tsx
