+++ src/components/app/app.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import App from "./app.jsx";
+import App from "./app";
 import {Provider} from "react-redux";
 import configureStore from "redux-mock-store";
 import {ALL_GENRES, SHOWED_MOVIES_DEFAULT} from "../../utils/constants";
