+++ src/components/big-movie-player/big-movie-player.tsx
