+++ src/components/genres-list/genres-list.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import {GenresList} from "./genres-list.jsx";
+import {GenresList} from "./genres-list";
 import {ALL_GENRES, SHOWED_MOVIES_DEFAULT} from "../../utils/constants";
 
 const films = [
