+++ src/components/main/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import GenresList from "../genres-list/genres-list.jsx";
-import ShowMoreButton from "../show-more-button/show-more-button.jsx";
+import GenresList from "../genres-list/genres-list";
+import ShowMoreButton from "../show-more-button/show-more-button";
 import {connect} from "react-redux";
 import {getPromoMovie} from "../../reducer/data/selectors.js";
 import {getAuthorizationStatus} from "../../reducer/user/selectors.js";
