+++ src/components/movie-card/movie-card.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import MovieCard from "./movie-card.jsx";
+import MovieCard from "./movie-card";
 
 Enzyme.configure({
   adapter: new Adapter()
