+++ src/components/movie-player/movie-player.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import MoviePlayer from "./movie-player.jsx";
+import MoviePlayer from "./movie-player";
 
 Enzyme.configure({
   adapter: new Adapter()
