+++ src/components/movies-list/movies-list.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {mount} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import MoviesList from "./movies-list.jsx";
+import MoviesList from "./movies-list";
 
 Enzyme.configure({
   adapter: new Adapter()
