+++ src/components/movies-list/movies-list.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import MoviesList from "./movies-list.jsx";
+import MoviesList from "./movies-list";
 
 const films = [
   {
