+++ src/components/my-list/my-list.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import MoviesList from "../movies-list/movies-list.jsx";
-import withActiveMovieCard from "../../hocs/with-active-movie-card/with-active-movie-card.jsx";
+import MoviesList from "../movies-list/movies-list";
+import withActiveMovieCard from "../../hocs/with-active-movie-card/with-active-movie-card";
 import PropTypes from "prop-types";
 import {Link} from "react-router-dom";
 import {AppRoute} from "../../utils/constants";
