+++ src/components/show-more-button/show-more-button.e2e.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, {mount} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import {ShowMoreButton} from "./show-more-button.jsx";
+import {ShowMoreButton} from "./show-more-button";
 import {SHOWED_MOVIES_DEFAULT} from "../../utils/constants.js";
 import {ALL_GENRES} from "../../utils/constants.js";
