+++ src/components/sign-in/sign-in.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import SignIn from "./sign-in.jsx";
+import SignIn from "./sign-in";
 import {MemoryRouter} from "react-router-dom";
 
 it(`SignIn component render correctly`, () => {
