+++ src/components/similar-movies/similar-movies.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import SimilarMovies from "./similar-movies.jsx";
+import SimilarMovies from "./similar-movies";
 
 const movie = {
   name: `Movie name`,
