+++ src/hocs/with-active-movie-card/with-active-movie-card.tsx
