+++ src/hocs/with-player/with-player.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import withPlayer from "./with-player.jsx";
+import withPlayer from "./with-player";
 
 const movie = {
   name: `Movie name`,
