+++ src/hocs/with-player/with-player.tsx
